# UCP/UCS/TEST: Optimize ptr_map lookup and small lookup fixes