# Why
- Eliminate some branches, since ptr_map values cannot be NULL. This affects all flows which use request/endpoint id: rdnv, sw rma/atomics, eager-sync, stream, active message with reply-ep.
- Fix lookup in ATS, atomic reply, and eager-sync-ack handlers
- Fix potential NULL access in ucp_worker_extract_request_by_id()

# How
- Return status from ptr_map lookup, instead of a pointer
- Checking the status code allows the compiler to eliminate branches

Before the PR (after adding the missing NULL check in ucp_eager_sync_ack_handler):
<table>
<tr>
<td>
Before: <b>0.154 usec</b>
</td>
<td>
After: <b>0.144 usec</b>
</td>
</tr>
<tr>
<td style="align:top">

```
ucp_eager_sync_ack_handler
  2.11 │       push   %r14
  2.50 │       push   %r13
  0.25 │       push   %r12
  5.00 │       push   %rbp
  2.06 │       push   %rbx
 25.50 │       mov    0x58(%rdi),%rax
 15.01 │       mov    0x1c8(%rax),%eax
  2.80 │       test   %eax,%eax
       │     ↓ jne    40
  0.39 │       mov    (%rsi),%rax
       │       test   $0x1,%al
       │     ↓ jne    50
       │       mov    %rax,%rdi
  6.47 │ 20:   test   %rdi,%rdi  <--- check for NULL
       │     ↓ je     31
  0.88 │       xor    %edx,%edx
       │       mov    $0x20,%esi
 27.37 │     → callq  ucp_tag_eager_sync_completion@plt
```
<td style="align:top">

```
ucp_eager_sync_ack_handler 
  1.69 │      push   %r14
  0.27 │      mov    %rdi,%rdx
  0.35 │      push   %r13
 22.63 │      push   %r12
  3.19 │      push   %rbp
       │      push   %rbx
 29.10 │      mov    0x58(%rdi),%rax
 19.25 │      mov    0x1c8(%rax),%eax
  3.11 │      test   %eax,%eax
       │    ↓ jne    40
  0.09 │      mov    (%rsi),%rax
       │      test   $0x1,%al
       │    ↓ jne    50
       │      mov    %rax,%rdi
  6.48 │23:   xor    %edx,%edx
       │      mov    $0x20,%esi
  0.18 │    → callq  ucp_tag_eager_sync_completion@plt
```
</td>
</td>
</tr>
</table>

Before:
```
UCX_TLS=shm mpirun -n 2 --report-bindings ./src/tools/pe/ucx_perftest -t tag_sync_bw -n 10000000000
MCW rank 0 bound to socket 0[core 0[hwt 0]]: [B/././././././././././.][./././././././././././.]
MCW rank 1 bound to socket 0[core 1[hwt 0]]: [./B/./././././././././.][./././././././././././.]
+--------------+--------------+-----------------------------+---------------------+-----------------------+
|              |              |      overhead (usec)        |   bandwidth (MB/s)  |  message rate (msg/s) |
+--------------+--------------+---------+---------+---------+----------+----------+-----------+-----------+
|    Stage     | # iterations | typical | average | overall |  average |  overall |  average  |  overall  |
+--------------+--------------+---------+---------+---------+----------+----------+-----------+-----------+
[thread 0]           6329913     0.127     0.158     0.158       48.29      48.29     6329881     6329881
[thread 0]          12942404     0.119     0.151     0.155       50.45      49.37     6612477     6471179
[thread 0]          19441151     0.109     0.154     0.154       49.58      49.44     6498714     6480357
[thread 0]          25983116     0.123     0.153     0.154       49.91      49.56     6541946     6495755
[thread 0]          32481133     0.130     0.154     0.154       49.58      49.56     6497997     6496203
[thread 0]          39045712     0.108     0.152     0.154       50.08      49.65     6564552     6507595
[thread 0]          45560136     0.125     0.154     0.154       49.70      49.66     6514393     6508566
```

After:
```
UCX_TLS=shm mpirun -n 2 --report-bindings ./src/tools/perf/ucx_perftest -t tag_sync_bw -n 10000000000
MCW rank 0 bound to socket 0[core 0[hwt 0]]: [B/././././././././././.][./././././././././././.]
MCW rank 1 bound to socket 0[core 1[hwt 0]]: [./B/./././././././././.][./././././././././././.]
+--------------+--------------+-----------------------------+---------------------+-----------------------+
|              |              |      overhead (usec)        |   bandwidth (MB/s)  |  message rate (msg/s) |
+--------------+--------------+---------+---------+---------+----------+----------+-----------+-----------+
|    Stage     | # iterations | typical | average | overall |  average |  overall |  average  |  overall  |
+--------------+--------------+---------+---------+---------+----------+----------+-----------+-----------+
[thread 0]           6960019     0.119     0.144     0.144       53.10      53.10     6959991     6959991
[thread 0]          13905029     0.117     0.144     0.144       52.99      53.04     6944988     6952490
[thread 0]          20847852     0.110     0.144     0.144       52.97      53.02     6942803     6949261
[thread 0]          27788395     0.115     0.144     0.144       52.95      53.00     6940508     6947073
[thread 0]          34768852     0.120     0.143     0.144       53.26      53.05     6980442     6953747
[thread 0]          41728976     0.115     0.144     0.144       53.10      53.06     6960104     6954806
```

Test system:
- Intel(R) Xeon(R) CPU E5-2650 v4 @ 2.20GHz
- gcc (GCC) 4.8.5 20150623 (Red Hat 4.8.5-36)