@hoopoepg 2021-01-17
it seems failure is relevant:
```
[2021-01-16T22:08:16.092Z] cc1: warnings being treated as errors
[2021-01-16T22:08:16.092Z] /scrap/jenkins/workspace/ucx/contrib/../src/uct/tcp/tcp_ep.c: In function ‘uct_tcp_ep_ptr_map_retrieve’:
[2021-01-16T22:08:16.092Z] /scrap/jenkins/workspace/ucx/contrib/../src/ucs/datastruct/ptr_map.inl:139: error: dereferencing pointer ‘ep.48’ does break strict-aliasing rules
[2021-01-16T22:08:16.092Z] /scrap/jenkins/workspace/ucx/contrib/../src/ucs/datastruct/ptr_map.inl:130: error: dereferencing pointer ‘ep.48’ does break strict-aliasing rules
[2021-01-16T22:08:16.092Z] /scrap/jenkins/workspace/ucx/contrib/../src/uct/tcp/tcp_ep.c:194: note: initialized from here
```